feat: 完成 NewShop 电商系统核心功能
后端模块:
- 路由注册与服务器启动
- 商品模型与仓库 (Product, Category, Brand, SKU)
- 商品 Service 与 Handler
- 购物车模块 (CartItem, Repository, Service, Handler)
- 订单模型与仓库 (Order, OrderItem, OrderLog)
- 订单服务与 Handler
- 支付宝支付集成 (smartwalle/alipay SDK)
- 预售模块 (Presale, PresaleOrder)
- 优惠券模块 (Coupon, UserCoupon)
- 积分签到模块 (PointsRecord, CheckIn)
- 会员等级模块 (MemberLevel, MemberExperience)
- 管理员认证 (Admin CRUD)
- 商品管理 API (商品/分类/品牌管理)
- 订单管理 API (发货/退款处理)
- 数据统计 API (销售/用户/商品统计)

前端模块:
- React + Vite + shadcn/ui + Tailwind CSS 项目初始化
- 前端布局与通用组件
- 用户认证页面 (登录/注册/找回密码)
- 商品展示页面 (首页/列表/详情/搜索)
- 购物车与订单页面 (购物车/结算/订单列表/订单详情)

+++ services/api/internal/service/coupon_service.go
@@ -208,7 +208,7 @@ func (s *CouponService) ValidateCoupon(ctx context.Context, userID, userCouponID
 	}
 
 	// 验证归属
-	if userCoupon.UserID != userID {
+	if userCoupon.UserID != uint64(userID) {
 		return ErrUserCouponNotFound
 	}
